requirements.txt
pyserial
webiopi
psutil